feat(chat): implement Redis caching and performance optimizations
Add Redis-based caching service for chat responses to improve performance
Optimize database queries with selectinload and limit chat history
Add performance monitoring decorator and health check endpoint
Implement streaming responses and background task processing
Update frontend with message virtualization and optimistic updates
 miner integration fixed
 onbarding not saving right values fixed

+++ alembic/versions/2bf43cdd4660_initial_migration_with_all_models.py
@@ -36,7 +36,7 @@ def upgrade() -> None:
     sa.Column('hashed_password', sa.String(length=255), nullable=False),
     sa.Column('is_active', sa.Boolean(), nullable=False),
     sa.Column('machine_name', sa.String(length=100), nullable=True),
-    sa.Column('contradiction_tolerance', sa.Integer(), nullable=True),
+    sa.Column('contradiction_tolerance', sa.Float(), nullable=True),
     sa.Column('belief_sensitivity', sa.String(length=50), nullable=True),
     sa.Column('salience_decay_speed', sa.String(length=50), nullable=True),
     sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
